fix: expose semicolons, remove generation warnings
This commit exposes semicolon terminators as anonymous nodes that can be
queried, and removes the following warnings upon generating the parser:

Warning: rule _pair_operator is just a `seq` or `choice` rule with a single element. This is unnecessary.
Warning: rule _lazy_or_operator is just a `seq` or `choice` rule with a single element. This is unnecessary.
Warning: rule _lazy_and_operator is just a `seq` or `choice` rule with a single element. This is unnecessary.
Warning: rule _pipe_right_operator is just a `seq` or `choice` rule with a single element. This is unnecessary.
Warning: rule _pipe_left_operator is just a `seq` or `choice` rule with a single element. This is unnecessary.
Warning: rule _rational_operator is just a `seq` or `choice` rule with a single element. This is unnecessary.
Warning: rule _tilde_operator is just a `seq` or `choice` rule with a single element. This is unnecessary.